requirements.txt
defusedxml==0.4.1
dj-database-url==0.3.0
Django==1.8.5
django-allauth==0.24.1
django-cors-headers==1.1.0
django-environ==0.4.0
django-rest-auth==0.6.0
django-taggit==0.18.0
django-taggit-serializer==0.1.5
djangorestframework==3.3.2
environ==1.0
httpsig==1.1.2
Markdown==2.6.5
oauthlib==1.0.3
Pillow==3.0.0
psycopg2==2.6.1
pycrypto==2.6.1
PyJWT==1.3.0
python3-openid==3.0.9
pytz==2015.7
requests==2.8.1
requests-oauthlib==0.6.0
six==1.10.0
stream-django==1.2.6
stream-python==2.3.4
wheel==0.24.0
